Move filter utils to factory
